fixing more issues when compiling for C++20

+++ tests/sim_speed/sc_main.cpp
@@ -5,7 +5,7 @@
 using namespace sc_core;
 
 int sc_main(int argc, char* argv[]) {
-    const uint64_t NS_VAL = 100000000;
+    const double NS_VAL = 100000000;
 
     sc_clock clk("clk", 1, SC_NS);
     sc_time run_time(NS_VAL, SC_NS);
